+++ task_checker_front/src/components/FormModal.vue
@@ -1,9 +1,9 @@
 <script setup>
 import { computed } from 'vue'
-import GenreBody from './GenreBody.vue'
-import TaskBody from './TaskBody.vue'
-import DetailBody from './DetailBody.vue'
-import EditBody from './EditBody.vue'
+import GenreBody from './FormModalBody/GenreBody.vue'
+import TaskBody from './FormModalBody/TaskBody.vue'
+import DetailBody from './FormModalBody/DetailBody.vue'
+import EditBody from './FormModalBody/EditBody.vue'
 
 const props = defineProps({
   body: String,
