+++ task_checker_front/src/components/FormModalBody/EditBody.vue
@@ -1,7 +1,7 @@
 <script setup>
-import Select from './Select.vue'
+import Select from '../Select.vue'
 import { ref } from 'vue'
-import { useTaskStore } from '../stores/taskStore';
+import { useTaskStore } from '../../stores/taskStore';
 
 const taskStore = useTaskStore();
